refactor(0164): split toctou.go along its three subjects
The file had grown to hold three unrelated things: path resolution
(ResolveAbsPathForCheck, DeepestExistingAncestor, ResolvePathForCheck,
ResolveAllForCheck), the judgement of what to check (ClassifyCheckTarget,
CollectPermissionCheckDirs and the two enums it dispatches on), and the
TOCTOU check itself (TOCTOUViolation, TOCTOUCheckResult,
RunTOCTOUPermissionCheck). Only the third is about TOCTOU; the first two
landed here because this branch added them as its front end. Every other
file in the package is named for one subject, so this one was the
exception.

Split into path_resolution.go, check_targets.go and toctou.go, with the
tests following the same three-way division. No identifier and no body
changed -- the same 23 test functions run, verified by diffing the
function lists before and after.

The three helpers shared by two of the test files (skipIfRoot,
blockedPathUnder, newBufferLogger) go to helpers_test.go rather than
test_helpers.go: they use only public APIs, so the test-organization
guide's decision tree puts them outside Classification B, and a
non-_test.go file loses the gosec/mnd exclusions that the chmod literals
in blockedPathUnder need.

The 0164 architecture and plan documents follow the new file names,
including the AC evidence table, which points at the moved
ResolvePathForCheck and ResolveAllForCheck tests.
